Refactor drawer components
