+++ src/components/drawers/commonDrawer/commonDrawer.scss
