+++ src/components/drawers/index.ts
@@ -1,0 +1,3 @@
+export * from './commonDrawer';
+export * from './exports';
+export * from './recommendations';
