+++ src/components/drawers/recommendations/recommendationsContent.tsx
