+++ src/routes/views/details/ocpDetails/detailsHeader.tsx
@@ -5,7 +5,7 @@ import { getProvidersQuery } from 'api/queries/providersQuery';
 import type { OcpReport } from 'api/reports/ocpReports';
 import { TagPathsType } from 'api/tags/tag';
 import type { AxiosError } from 'axios';
-import { ExportsLink } from 'components/exports';
+import { ExportsLink } from 'components/drawers';
 import messages from 'locales/messages';
 import React from 'react';
 import type { WrappedComponentProps } from 'react-intl';
